Mantis 1057 - Returns structural analysis results tolerance back to default values.
The cause of the structural analysis error was a bad equilibrium check in WBFL distributed loads. That bug has been in FEM2D since it's inception. The tolerances kept getting set to larger and larger values to work around the problem. Now that the problem seems to be fixed, the tolerances have been set back to default values.